Use html video tag for streaming
the playerTemplate.html has been modified to use a VideoInfo struct to
fill in the video path.
Using player=1 in the query string of the index endpoint now returns an
html file generated with the streamListTemplate.html template instead of
listTemplate.html. This template is diferent because the links point to
the playerTemplate instead of the raw video file. this should be used in
browsers that support the html5 video tag but do not support streaming
raw video files like the PS4.

The GODEBUG environment variable now sets the Server.go debug flag, instead of
setting it manually in the source code.

The server no longer crashes when the computer is not connected to the
internet.

Better documentation for exported methods.